Add ImGui.MenuBar(bool)
This enables or disables the menu bar and all calls to lua. This effectively disables everything **except** the trampoline required for imgui.

To properly disable imgui define `imgui_disabled` variable in the global scope and set it to true. All imgui calls will be replaced with stubs and the trampoline will never be setup making it safe for production.